third_party/blink/web_tests/http/tests/websocket/send-arraybufferview-expected.txt

WebSocket: Send ArrayBufferViews.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.send(createSharedArrayBufferView()) threw exception TypeError: Failed to execute 'send' on 'WebSocket': The provided ArrayBufferView value must not be shared..
PASS PASS: Message #0.
PASS PASS: Message #1.
PASS PASS: Message #2.
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
